change class name.   add doc
